+++ InventoryTools/InventoryToolsPlugin.cs
@@ -76,13 +76,13 @@ namespace InventoryTools
             ICondition condition, IDataManager dataManager, IFramework framework, IGameGui gameGui,
             IGameInteropProvider gameInteropProvider, IKeyState keyState, IObjectTable objectTable, ITargetManager targetManager, ITextureProvider textureProvider,
             IToastGui toastGui, IContextMenu contextMenu, ITitleScreenMenu titleScreenMenu,
-            IGameInventory gameInventory) : base(pluginInterface,
+            IGameInventory gameInventory, IPlayerState playerState) : base(pluginInterface,
             pluginLog, addonLifecycle, chatGui, clientState, commandManager,
             condition, dataManager, framework, gameGui,
             gameInteropProvider, keyState, objectTable,
             targetManager, textureProvider,
             toastGui, contextMenu, titleScreenMenu,
-            gameInventory)
+            gameInventory, playerState)
         {
             Stopwatch loadConfigStopwatch = new Stopwatch();
             loadConfigStopwatch.Start();