+++ InventoryTools/Logic/Columns/Abstract/ColumnCategory.cs
@@ -5,6 +5,7 @@ public enum ColumnCategory
     Unsorted,
     Basic,
     Stats,
+    Desynthesis,
     Inventory,
     Crafting,
     Market,